requirements.txt
aiogram==3.13.1
asyncpg
aiohttp==3.10.5
python-dotenv==1.0.1
fastapi
uvicorn[standard]
pydantic
